third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Document-importNode-expected.txt

This is a testharness.js-based test.
[FAIL] importNode should clone using target's registry if non-null, including when it's not the global registry
  assert_false: expected false got true
[FAIL] importNode should clone a template content using a specified scoped registry
  assert_equals: expected object "[object CustomElementRegistry]" but got object "[object CustomElementRegistry]"
Harness: the test ran to completion.
